feat: change table structure
