+++ front-end/src/components/Table/BitcoinTable/styles.js
