+++ front-end/src/components/Table/HybridTable/styles.js
