+++ front-end/src/components/Table/styles.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const TableContainer = styled.div`
+  display: flex;
+  width: 100%;
+  height: 70%;
+  align-items: center;
+  justify-content: center;
+  //margin: 25px 0px;
+  padding: 10px;
+  background: #f8fafd;
+`;
